Add authorization to delete approval

+++ src/main/java/com/ercanbeyen/bloggingplatform/service/impl/ApprovalServiceImpl.java
@@ -6,13 +6,16 @@ import com.ercanbeyen.bloggingplatform.dto.ApprovalDto;
 import com.ercanbeyen.bloggingplatform.dto.converter.ApprovalDtoConverter;
 import com.ercanbeyen.bloggingplatform.dto.request.create.CreateApprovalRequest;
 import com.ercanbeyen.bloggingplatform.entity.Approval;
+import com.ercanbeyen.bloggingplatform.entity.Author;
 import com.ercanbeyen.bloggingplatform.entity.Ticket;
 import com.ercanbeyen.bloggingplatform.exception.data.DataConflict;
+import com.ercanbeyen.bloggingplatform.exception.data.DataForbidden;
 import com.ercanbeyen.bloggingplatform.exception.data.DataNotFound;
 import com.ercanbeyen.bloggingplatform.mapper.ApprovalMapper;
 import com.ercanbeyen.bloggingplatform.service.ApprovalService;
 import com.ercanbeyen.bloggingplatform.service.AuthorService;
 import com.ercanbeyen.bloggingplatform.service.TicketService;
+import com.ercanbeyen.bloggingplatform.util.SecurityUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -51,7 +54,15 @@ public class ApprovalServiceImpl implements ApprovalService {
 
     @Override
     public String deleteApproval(String id) {
+        Author author = SecurityUtil.getLoggedInAuthor();
+        Approval approval = approvalMapper.findApprovalById(id);
+
+        if (!approval.getAuthorId().equals(author.getId())) {
+            throw new DataForbidden(ResponseMessage.NOT_AUTHORIZED);
+        }
+
         approvalMapper.deleteApprovalById(id);
+
         return String.format(ResponseMessage.SUCCESS, EntityName.APPROVAL, id, ResponseMessage.Operation.DELETED);
     }
